modified:   app/Http/Controllers/Admin/CheatingIncidentController.php 	modified:   app/Http/Controllers/Admin/CourseController.php 	modified:   app/Http/Controllers/Admin/DocumentationController.php 	modified:   app/Http/Controllers/Admin/ForumCategoryController.php 	modified:   app/Http/Controllers/Admin/InformationCardController.php 	modified:   app/Http/Controllers/Admin/MaterialController.php 	modified:   app/Http/Controllers/Admin/QuestionBankController.php 	modified:   app/Http/Controllers/Admin/QuestionController.php 	modified:   app/Http/Controllers/Admin/SchoolClassController.php 	modified:   app/Http/Controllers/Admin/SettingsController.php 	modified:   app/Http/Controllers/Admin/UserController.php 	modified:   app/Http/Controllers/CertificateController.php 	modified:   app/Http/Controllers/ForumController.php 	modified:   app/Http/Controllers/GuestExamController.php 	modified:   app/Http/Controllers/Guru/CourseController.php 	modified:   app/Http/Controllers/Guru/InformationCardController.php 	modified:   app/Http/Controllers/Guru/MaterialController.php 	modified:   app/Http/Controllers/Guru/QuestionBankController.php 	modified:   app/Http/Controllers/Guru/QuestionController.php 	modified:   app/Http/Controllers/MaterialCommentController.php 	modified:   app/Http/Controllers/OfflineExamController.php 	modified:   app/Http/Controllers/Siswa/AssignmentController.php 	modified:   app/Http/Controllers/Siswa/CertificateController.php 	modified:   app/Http/Controllers/Siswa/GradeController.php 	modified:   app/Models/Assignment.php 	modified:   app/Models/AssignmentSubmission.php 	modified:   app/Models/Certificate.php 	modified:   app/Models/Course.php 	modified:   app/Models/CourseGradeWeight.php 	modified:   app/Models/CourseGroup.php 	modified:   app/Models/Enrollment.php 	modified:   app/Models/Exam.php 	modified:   app/Models/ExamAttempt.php 	modified:   app/Models/User.php 	modified:   app/Policies/CoursePolicy.php 	modified:   app/Policies/EnrollmentPolicy.php 	new file:   app/Policies/QuestionBankPolicy.php 	modified:   composer.json 	modified:   composer.lock 	new file:   config/purifier.php 	modified:   resources/css/app.css 	modified:   resources/views/admin/error-logs/index.blade.php 	modified:   resources/views/admin/exams/show.blade.php 	modified:   resources/views/admin/question-bank/create.blade.php 	modified:   resources/views/admin/question-bank/edit.blade.php 	modified:   resources/views/admin/question-bank/index.blade.php 	modified:   resources/views/admin/question-bank/pdf-export.blade.php 	modified:   resources/views/admin/question-bank/show.blade.php 	modified:   resources/views/admin/question-bank/statistics.blade.php 	modified:   resources/views/admin/questions/create.blade.php 	modified:   resources/views/admin/questions/edit.blade.php 	modified:   resources/views/admin/questions/index.blade.php 	modified:   resources/views/components/information-cards.blade.php 	modified:   resources/views/guest/exams/review.blade.php 	modified:   resources/views/guest/exams/take.blade.php 	modified:   resources/views/guru/exams/review-essays.blade.php 	modified:   resources/views/guru/exams/show.blade.php 	modified:   resources/views/guru/question-bank/create.blade.php 	modified:   resources/views/guru/question-bank/edit.blade.php 	modified:   resources/views/guru/question-bank/index.blade.php 	modified:   resources/views/guru/question-bank/show.blade.php 	modified:   resources/views/guru/questions/create.blade.php 	modified:   resources/views/guru/questions/edit.blade.php 	modified:   resources/views/guru/questions/index.blade.php 	modified:   resources/views/landing.blade.php 	modified:   resources/views/offline/exams/take.blade.php 	modified:   resources/views/profile/edit.blade.php 	modified:   resources/views/siswa/exams/review.blade.php 	modified:   resources/views/siswa/exams/take.blade.php 	modified:   routes/certificates.php 	modified:   routes/offline.php 	modified:   routes/web.php

+++ app/Http/Controllers/Admin/ForumCategoryController.php
@@ -93,6 +93,11 @@ class ForumCategoryController extends Controller
      */
     public function destroy(ForumCategory $forumCategory)
     {
+        // Bug #30: Check for existing threads before deleting
+        if ($forumCategory->threads()->count() > 0) {
+            return back()->with('error', __('Tidak dapat menghapus kategori yang masih memiliki thread.'));
+        }
+
         $forumCategory->delete();
 
         return redirect()->route('admin.forum-categories.index')
